+++ CHANGELOG.md
@@ -58,9 +58,14 @@ rewrite methods which using GSON.
 
 ## 1.9.2
 
-* Add extensions for `GSONBuilder` for implementation correct [IInputObject](src/main/kotlin/com/github/insanusmokrassar/IObjectKRealisations/Extensions.kt#72)
+* Add extensions for `GSONBuilder` for implementation correct [IInputObject](src/main/kotlin/com/github/insanusmokrassar/IObjectKRealisations/Extensions.kt)
 
 ## 1.9.3
 
 * Remove redundant `keys` from [PropertiesIObject](src/main/kotlin/com/github/insanusmokrassar/IObjectKRealisations/PropertiesIObject.kt)
 * Fix constructors work in [PropertiesIObject](src/main/kotlin/com/github/insanusmokrassar/IObjectKRealisations/PropertiesIObject.kt)
+
+## 1.9.4
+
+* Update dependencies
+* Fix [IInputString<String, in Any>#toObject](src/main/kotlin/com/github/insanusmokrassar/IObjectKRealisations/Extensions.kt#25)